+++ lib/pyth_sui_dart.dart
@@ -1,7 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library;
 
+export 'src/price_feed.dart';
 export 'src/sui_price_service_connection.dart';
 export 'src/client.dart';